+++ src/Library/Mensaje/Message.cs
@@ -15,7 +15,7 @@ namespace Bot
         /// </summary>
         /// <param name="id">Id del usuario que envía el mensaje</param>
         /// <param name="message">Contenido del mensaje</param>
-        public Message(int id, string message)
+        public Message(long id, string message)
         {
             this.UserId = id;
             this.Text = message;
@@ -23,7 +23,7 @@ namespace Bot
         /// <summary>
         /// Id del usuario que esta charlando
         /// </summary>
-        public int UserId { get; set; }
+        public long UserId { get; set; }
 
         /// <summary>
         /// Mensaje del usuario
